+++ src/collective/solr/events.py
@@ -1,8 +1,10 @@
+# -*- coding: utf-8 -*-
 from Products.CMFCore.utils import getToolByName
 from Products.CMFDefault.permissions import ModifyPortalContent
 
 from Products.CMFCore.CMFCatalogAware import CMFCatalogAware
 from Products.Archetypes.CatalogMultiplex import CatalogMultiplex
+
 
 def reorderedEvent(event):
     parent = event.object