+++ src/collective/solr/exportimport.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from persistent.interfaces import IPersistent
 from zope.component import queryUtility
 from Products.GenericSetup.utils import exportObjects
@@ -34,7 +35,7 @@ class SolrConfigXMLAdapter(XMLAdapterBase):
         self.context.commit_within = 0
         self.context.index_timeout = 0
         self.context.search_timeout = 0
-        self.context.max_results = 0
+        self.context.max_results = 10000000
         self.context.required = []
         self.context.search_pattern = ''
         self.context.facets = []
@@ -46,6 +47,7 @@ class SolrConfigXMLAdapter(XMLAdapterBase):
         self.context.highlight_formatter_pre = ''
         self.context.highlight_formatter_post = ''
         self.context.highlight_fragsize = 0
+        self.context.levenshtein_distance = 0
 
     def _initProperties(self, node):
         elems = node.getElementsByTagName('connection')
@@ -132,6 +134,9 @@ class SolrConfigXMLAdapter(XMLAdapterBase):
                     for elem in child.getElementsByTagName('parameter'):
                         value.append(elem.getAttribute('name'))
                     self.context.field_list = tuple(map(str, value))
+                elif child.nodeName == 'levenshtein_distance':
+                    value = float(str(child.getAttribute('value')))
+                    self.context.levenshtein_distance = value
 
     def _createNode(self, name, value):
         node = self._doc.createElement(name)
@@ -191,6 +196,8 @@ class SolrConfigXMLAdapter(XMLAdapterBase):
         append(create('highlight_fragsize', str(self.context.highlight_fragsize)))
         field_list = self._doc.createElement('field-list')
         append(field_list)
+        append(create('levenshtein_distance',
+            str(self.context.levenshtein_distance)))
         for name in self.context.field_list:
             param = self._doc.createElement('parameter')
             param.setAttribute('name', name)