+++ src/collective/solr/lingua.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 try:
     from Products.LinguaPlone.catalog import languageFilter
     languageFilter      # keep pyflakes happy