+++ src/collective/solr/solr.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # http://svn.apache.org/repos/asf/lucene/solr/trunk/client/python/solr.py
 
 # Licensed to the Apache Software Foundation (ASF) under one or more
@@ -27,7 +28,6 @@
 # c.delete('123')
 # c.commit()
 
-import sys
 import httplib
 import socket
 from xml.etree.cElementTree import fromstring
@@ -52,7 +52,8 @@ class SolrException(Exception):
 
     def __repr__(self):
         return 'HTTP code=%s, Reason=%s, body=%s' % (
-                    self.httpcode, self.reason, self.body)
+            self.httpcode, self.reason, self.body
+        )
 
     def __str__(self):
         return 'HTTP code=%s, reason=%s' % (self.httpcode, self.reason)
@@ -63,7 +64,7 @@ class SolrConnection:
     def __init__(self, host='localhost:8983', solrBase='/solr',
                  persistent=True, postHeaders={}, timeout=None):
         self.host = host
-        self.solrBase = solrBase
+        self.solrBase = str(solrBase)
         self.persistent = persistent
         self.reconnects = 0
         self.encoder = codecs.getencoder('utf-8')
@@ -76,16 +77,24 @@ class SolrConnection:
         self.xmlheaders = {'Content-Type': 'text/xml; charset=utf-8'}
         self.xmlheaders.update(postHeaders)
         if not self.persistent:
-            self.xmlheaders['Connection']='close'
-        self.formheaders = {'Content-Type':
-            'application/x-www-form-urlencoded; charset=utf-8'}
+            self.xmlheaders['Connection'] = 'close'
+        self.formheaders = {
+            'Content-Type': 'application/x-www-form-urlencoded; charset=utf-8'
+        }
         if not self.persistent:
-            self.formheaders['Connection']='close'
+            self.formheaders['Connection'] = 'close'
 
     def __str__(self):
-        return ('SolrConnection{host=%s, solrBase=%s, persistent=%s, '
-            'postHeaders=%s, reconnects=%s}' % (self.host, self.solrBase,
-            self.persistent, self.xmlheaders, self.reconnects))
+        return (
+            'SolrConnection{host=%s, solrBase=%s, persistent=%s, '
+            'postHeaders=%s, reconnects=%s}' % (
+                self.host,
+                self.solrBase,
+                self.persistent,
+                self.xmlheaders,
+                self.reconnects
+            )
+        )
 
     def __reconnect(self):
         self.reconnects += 1
@@ -116,8 +125,10 @@ class SolrConnection:
         try:
             self.conn.request('POST', url, body, headers)
             return self.__errcheck(self.conn.getresponse())
-        except (socket.error, httplib.CannotSendRequest,
-            httplib.ResponseNotReady, httplib.BadStatusLine):
+        except (
+            socket.error, httplib.CannotSendRequest,
+            httplib.ResponseNotReady, httplib.BadStatusLine
+        ):
             # Reconnect in case the connection was broken from the server
             # going down, the server timing out our persistent connection, or
             # another network failure. Also catch httplib.CannotSendRequest,
@@ -145,15 +156,19 @@ class SolrConnection:
             except (SolrException, socket.error):
                 logger.exception('exception during request %r', request)
             count += len(request)
-        logger.debug('flushed out %d bytes in %d requests',
-            count, len(self.xmlbody))
+        logger.debug(
+            'flushed out %d bytes in %d requests',
+            count, len(self.xmlbody)
+        )
         del self.xmlbody[:]
         return responses
 
     def doSendXML(self, request):
         try:
-            rsp = self.doPost(self.solrBase+'/update', request,
-                self.xmlheaders)
+            rsp = self.doPost(
+                self.solrBase+'/update', request,
+                self.xmlheaders
+            )
             data = rsp.read()
         finally:
             if not self.persistent:
@@ -209,7 +224,7 @@ class SolrConnection:
                     self.escapeKey(f), boost_values[f])
             else:
                 tmpl = '<field name="%s">%%s</field>' % self.escapeKey(f)
-            if isinstance(v, (list, tuple)): # multi-valued
+            if isinstance(v, (list, tuple)):  # multi-valued
                 for value in v:
                     lst.append(tmpl % self.escapeVal(value))
             else:
@@ -220,10 +235,12 @@ class SolrConnection:
         return self.doUpdateXML(xstr)
 
     def commit(self, waitFlush=True, waitSearcher=True, optimize=False):
-        data = {'committype': optimize and 'optimize' or 'commit',
-                'nowait': not waitSearcher and ' waitSearcher="false"' or '',
-                'noflush': not waitFlush and not waitSearcher and \
-                    ' waitFlush="false"' or ''}
+        data = {
+            'committype': optimize and 'optimize' or 'commit',
+            'nowait': not waitSearcher and ' waitSearcher="false"' or '',
+            'noflush': not waitFlush and not waitSearcher and
+            ' waitFlush="false"' or ''
+        }
         xstr = '<%(committype)s%(noflush)s%(nowait)s/>' % data
         self.doUpdateXML(xstr)
         return self.flush()
@@ -233,31 +250,39 @@ class SolrConnection:
         # for now we delay sending the xml until the commit (see above),
         # which is why we don't have to send anything to abort...
         # see http://issues.apache.org/jira/browse/SOLR-670
-        logger.debug('aborting %d requests: %r',
-            len(self.xmlbody), self.xmlbody)
+        logger.debug(
+            'aborting %d requests: %r',
+            len(self.xmlbody),
+            self.xmlbody
+        )
         del self.xmlbody[:]
 
     def search(self, **params):
         request = urllib.urlencode(params, doseq=True)
         logger.debug('sending request: %s' % request)
         try:
-            response = self.doPost('%s/select' % self.solrBase, request,
-                self.formheaders)
+            response = self.doPost(
+                '%s/select' % self.solrBase, request,
+                self.formheaders
+            )
         finally:
             if not self.persistent:
                 self.conn.close()
         return response
 
     def getSchema(self):
-        schema_urls = ('%s/admin/file/?file=schema.xml',        # solr 1.3
-                       '%s/admin/get-file.jsp?file=schema.xml') # solr 1.2
+        schema_urls = (
+            '%s/admin/file/?file=schema.xml',         # solr 1.3
+            '%s/admin/get-file.jsp?file=schema.xml')  # solr 1.2
         for url in schema_urls:
             logger.debug('getting schema from: %s', url % self.solrBase)
             try:
                 self.conn.request('GET', url % self.solrBase)
                 response = self.conn.getresponse()
-            except (socket.error, httplib.CannotSendRequest,
-                httplib.ResponseNotReady, httplib.BadStatusLine):
+            except (
+                socket.error, httplib.CannotSendRequest,
+                httplib.ResponseNotReady, httplib.BadStatusLine
+            ):
                 # see `doPost` method for more info about these exceptions
                 self.__reconnect()
                 self.conn.request('GET', url % self.solrBase)