+++ src/collective/solr/tests/test_exportimport.py
@@ -28,6 +28,7 @@ class SetupToolTests(SolrTestCase, TarballTester):
         config.slow_query_threshold = 2342
         config.effective_steps = 900
         config.exclude_user = True
+        config.levenshtein_distance = 0.2
 
     def testImportStep(self):
         profile = 'profile-collective.solr:default'
@@ -46,13 +47,14 @@ class SetupToolTests(SolrTestCase, TarballTester):
         self.assertEqual(config.commit_within, 1000)
         self.assertEqual(config.index_timeout, 0)
         self.assertEqual(config.search_timeout, 0)
-        self.assertEqual(config.max_results, 0)
+        self.assertEqual(config.max_results, 10000000)
         self.assertEqual(config.required, ('SearchableText', ))
         self.assertEqual(config.facets, ('portal_type', 'review_state'))
         self.assertEqual(config.filter_queries, ('portal_type', ))
         self.assertEqual(config.slow_query_threshold, 0)
         self.assertEqual(config.effective_steps, 1)
         self.assertEqual(config.exclude_user, False)
+        self.assertEqual(config.levenshtein_distance, 0.0)
 
     def testExportStep(self):
         tool = self.portal.portal_setup
@@ -121,6 +123,7 @@ SOLR_XML = """\
     <highlight_formatter_post value="]"/>
     <highlight_fragsize value="100"/>
     <field-list/>
+    <levenshtein_distance value="0.2"/>
   </settings>
 </object>
 """