+++ src/collective/solr/tests/test_mangler.py
@@ -144,57 +144,57 @@ class PathManglerTests(TestCase):
 
     def testSimplePathQuery(self):
         keywords = mangle(path='/foo')
-        self.assertEqual(keywords, {'path_parents': '/foo'})
+        self.assertEqual(keywords, {'path_parents': '"\\/foo"'})
 
     def testSimplePathQueryAsDictionary(self):
         keywords = mangle(path=dict(query='/foo'))
-        self.assertEqual(keywords, {'path_parents': '/foo'})
+        self.assertEqual(keywords, {'path_parents': '"\\/foo"'})
 
     def testSimplePathQueryAsObject(self):
         keywords = mangle(path=Query(query='/foo'))
-        self.assertEqual(keywords, {'path_parents': '/foo'})
+        self.assertEqual(keywords, {'path_parents': '"\\/foo"'})
 
     def testPathQueryWithDepth(self):
         keywords = mangle(path=dict(query='/foo', depth=-1))
-        self.assertEqual(keywords, {'path_parents': '/foo'})
+        self.assertEqual(keywords, {'path_parents': '"\\/foo"'})
         keywords = mangle(path=dict(query='/foo', depth=0))
         self.assertEqual(keywords, {'path_parents': set([
-            '(+path_depth:[2 TO 2] AND +path_parents:/foo)'])})
+            '+(+path_depth:[2 TO 2] AND +path_parents:"\\/foo")'])})
         keywords = mangle(path=Query(query='/foo', depth=0))
         self.assertEqual(keywords, {'path_parents': set([
-            '(+path_depth:[2 TO 2] AND +path_parents:/foo)'])})
+            '+(+path_depth:[2 TO 2] AND +path_parents:"\\/foo")'])})
         keywords = mangle(path=dict(query='/foo', depth=2))
         self.assertEqual(keywords, {'path_parents': set([
-            '(+path_depth:[3 TO 4] AND +path_parents:/foo)'])})
+            '+(+path_depth:[3 TO 4] AND +path_parents:"\\/foo")'])})
         keywords = mangle(path=Query(query='/foo', depth=2))
         self.assertEqual(keywords, {'path_parents': set([
-            '(+path_depth:[3 TO 4] AND +path_parents:/foo)'])})
+            '+(+path_depth:[3 TO 4] AND +path_parents:"\\/foo")'])})
 
     def testMultiplePathQuery(self):
         keywords = mangle(path=['/foo', '/bar'])
-        self.assertEqual(keywords, {'path_parents': ['/foo', '/bar']})
+        self.assertEqual(keywords, {'path_parents': ['"\\/foo"', '"\\/bar"']})
         keywords = mangle(path=dict(query=['/foo', '/bar']))
-        self.assertEqual(keywords, {'path_parents': ['/foo', '/bar']})
+        self.assertEqual(keywords, {'path_parents': ['"\\/foo"', '"\\/bar"']})
         keywords = mangle(path=dict(query=['/foo', '/bar'], depth=-1))
-        self.assertEqual(keywords, {'path_parents': ['/foo', '/bar']})
+        self.assertEqual(keywords, {'path_parents': ['"\\/foo"', '"\\/bar"']})
         keywords = mangle(path=dict(query=['/foo', '/bar'], depth=0))
         self.assertEqual(keywords, {'path_parents': set([
-            '(+path_depth:[2 TO 2] AND +path_parents:/foo)',
-            '(+path_depth:[2 TO 2] AND +path_parents:/bar)'])})
+            '+(+path_depth:[2 TO 2] AND +path_parents:"\\/foo")',
+            '+(+path_depth:[2 TO 2] AND +path_parents:"\\/bar")'])})
         keywords = mangle(path=dict(query=['/foo', '/bar'], depth=1))
         self.assertEqual(keywords, {'path_parents': set([
-            '(+path_depth:[3 TO 3] AND +path_parents:/foo)',
-            '(+path_depth:[3 TO 3] AND +path_parents:/bar)'])})
+            '+(+path_depth:[3 TO 3] AND +path_parents:"\\/foo")',
+            '+(+path_depth:[3 TO 3] AND +path_parents:"\\/bar")'])})
         keywords = mangle(path=dict(query=['/a/b', '/c'], depth=-1))
-        self.assertEqual(keywords, {'path_parents': ['/a/b', '/c']})
+        self.assertEqual(keywords, {'path_parents': ['"\\/a\\/b"', '"\\/c"']})
         keywords = mangle(path=dict(query=['/a/b', '/c'], depth=0))
         self.assertEqual(keywords, {'path_parents': set([
-            '(+path_depth:[3 TO 3] AND +path_parents:/a/b)',
-            '(+path_depth:[2 TO 2] AND +path_parents:/c)'])})
+            '+(+path_depth:[3 TO 3] AND +path_parents:"\\/a\\/b")',
+            '+(+path_depth:[2 TO 2] AND +path_parents:"\\/c")'])})
         keywords = mangle(path=dict(query=['/a/b', '/c'], depth=2))
         self.assertEqual(keywords, {'path_parents': set([
-            '(+path_depth:[4 TO 5] AND +path_parents:/a/b)',
-            '(+path_depth:[3 TO 4] AND +path_parents:/c)'])})
+            '+(+path_depth:[4 TO 5] AND +path_parents:"\\/a\\/b")',
+            '+(+path_depth:[3 TO 4] AND +path_parents:"\\/c")'])})
 
 
 class QueryParameterTests(TestCase):